wallProfile ui element working added some comment and cleaned code
