--- app/src/main/res/layout/wall_graph.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<view class="fi.climbstationsolutions.climbstation.utils.WallGraph"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/wall_graph"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent" />
