--- app/src/main/res/values/wall_graph_attrs.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <declare-styleable name="WallGraph">
-        <attr name="profile"/>
-    </declare-styleable>
-</resources>
