+++ app/src/main/res/values/wall_profile_attrs.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <declare-styleable name="WallProfile">
+        <attr name="wallColor" format="color"/>
+        <attr name="wallFillColor" format="color"/>
+    </declare-styleable>
+</resources>
